+++ js/model3D/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/model3d
 
+## 0.14.16
+
+### Dependency updates
+
+- @gradio/upload@0.16.7
+- @gradio/client@1.15.2
+
 ## 0.14.15
 
 ### Dependency updates